Remove group from material and move renderChildren helper to vanilla
