+++ packages/material/src/layouts/index.ts
@@ -1,3 +1,2 @@
-export * from './group.layout';
 export * from './MaterialHorizontalLayout';
 export * from './MaterialVerticalLayout';
