+++ packages/vanilla/src/layouts/horizontal.layout.tsx
@@ -5,11 +5,10 @@ import {
   RankedTester,
   rankWith,
   registerStartupRenderer,
-  renderChildren,
   uiTypeIs,
 } from '@jsonforms/core';
 import { connect } from 'react-redux';
-import { mapStateToVanillaLayoutProps, VanillaRendererProps } from '../helpers';
+import { mapStateToVanillaLayoutProps, renderChildren, VanillaRendererProps } from '../helpers';
 
 /**
  * Default tester for a horizontal layout.
@@ -41,14 +40,7 @@ const HorizontalLayoutRenderer = (
       className={layoutClassName}
       visible={visible}
     >
-      {
-        renderChildren(
-          horizontalLayout,
-          schema,
-          childClassNames,
-          path
-        )
-      }
+      {renderChildren(horizontalLayout, schema, childClassNames, path)}
     </JsonFormsLayout>
   );
 };
